feat: use CauldronV2 instead CauldronV2MultiChain, rename in deploy files, change and rename CauldronCook script, rename variables in CauldronV2
