+++ deploy/CauldronV2.ts
@@ -3,7 +3,7 @@ import { DeployFunction } from "hardhat-deploy/types";
 import { ethers, network } from "hardhat";
 import { ChainId, setDeploymentSupportedChains } from "../utilities";
 import { LothricFin } from "../test/constants";
-import { CauldronV2MultiChain } from "../typechain";
+import { CauldronV2 } from "../typechain";
 
 const ParametersPerChain = {
   [ChainId.Localhost]: {
@@ -23,20 +23,20 @@ const deployFunction: DeployFunction = async function (hre: HardhatRuntimeEnviro
   const nusd = (await deployments.get("NereusStableCoin")).address;
   const degenBox = (await deployments.get("DegenBox")).address;
 
-  const tx = await deploy("CauldronV2MultiChain", {
+  const tx = await deploy("CauldronV2", {
     from: deployer,
     args: [degenBox, nusd],
     log: true,
     deterministicDeployment: false,
   });
 
-  const CauldronV2MultiChain = await ethers.getContract<CauldronV2MultiChain>("CauldronV2MultiChain");
+  const CauldronV2 = await ethers.getContract<CauldronV2>("CauldronV2");
 
-  if ((await CauldronV2MultiChain.owner()) != parameters.owner && network.name !== "hardhat") {
-    await CauldronV2MultiChain.transferOwnership(parameters.owner, true, false);
+  if ((await CauldronV2.owner()) != parameters.owner && network.name !== "hardhat") {
+    await CauldronV2.transferOwnership(parameters.owner, true, false);
   }
 
-  await deployments.save("CauldronV2MultiChain", {
+  await deployments.save("CauldronV2", {
     abi: [],
     address: tx.address,
   });
@@ -46,5 +46,5 @@ export default deployFunction;
 
 setDeploymentSupportedChains(Object.keys(ParametersPerChain), deployFunction);
 
-deployFunction.tags = ["CauldronV2MultiChain"];
+deployFunction.tags = ["CauldronV2"];
 deployFunction.dependencies = ["DegenBox"];
